rename `base_widgets` package to `widget`
