+++ susapad/window/settings/component/close.py
@@ -1,10 +1,10 @@
 
 from PySide6 import QtCore
 
-from susapad import base_widgets as base
+from susapad import widget
 
 
-class CloseButton(base.BaseFloatingButton):
+class CloseButton(widget.BaseFloatingButton):
 
     def __init__(self, window, main_window):
         super().__init__(window, "Fechar", "Escape")
